% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas.R
\name{betas}
\alias{betas}
\title{Record the Probabilistic Event}
\usage{
betas(case_num = 0, model, concept, doc, topic_num = 0, math = 0)
}
\arguments{
\item{case_num}{The case number}

\item{model}{The model generated from ldamodel() function}

\item{concept}{The thematic concept (viz. 'school', 'family', 'work')}

\item{doc}{The document in question}

\item{topic_num}{The topic number from the reviewed topics list.}

\item{math}{The number that needs to be subtracted from the beta sum, based on the kwic concordance}
}
\value{
An artifact qualifying as the record of the probabilistic event
}
\description{
Record the Probabilistic Event
}
\examples{
output <- betas(case_num = 1, model = model, concept = 'work', doc = doc, topic_num = 3, math = 0)
}
