% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essay.R
\name{essay}
\alias{essay}
\title{Create a corpus object}
\usage{
essay(a = data_dir, b = file, d = 0)
}
\arguments{
\item{a}{The file directory holding the data set}

\item{b}{The file in question}

\item{d}{The case number}
}
\value{
The corpus object
}
\description{
Create a corpus object
}
\examples{
doc <- essay(data_dir, file, 1)
}
