% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesave.R
\name{filesave}
\alias{filesave}
\title{Save the Output}
\usage{
filesave(save_dir = in_quotes, file_name = in_quotes)
}
\arguments{
\item{save_dir}{The directory path.}

\item{file_name}{The name of the file (without extension or dot), in quotes.}
}
\value{
saved evidence in an RDS file to capture artifacts from the probabilistic event.
}
\description{
Save the Output
}
\examples{
filesave(data_dir, "case1")
}
