% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savet.R
\name{savet}
\alias{savet}
\title{This will save an instance of Gibbs topic models from the ldamodel function}
\usage{
savet(a = topics, b = data_dir, d = case)
}
\arguments{
\item{a}{topics object from ldamodel (from the general environment)}

\item{b}{The data directory}

\item{d}{The Case number}
}
\value{
Text number, Token Types, Sentence numbers, and Title. An abbreviated version of summary in quanteda
}
\description{
This will save an instance of Gibbs topic models from the ldamodel function
}
\examples{
mod <- ldamodel(features, 8)
model <- mod$gibbs
topics <- mod$words

savet(topics, data_dir, 1)
}
