% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdfm.R
\name{tmdfm}
\alias{tmdfm}
\title{Convert the corpus to prepare for ldatuning}
\usage{
tmdfm(doc = doc)
}
\arguments{
\item{doc}{Referring to the object created by the essay function}
}
\value{
Text number, Token Types, Sentence numbers, and Title. An abbreviated version of summary in quanteda
}
\description{
Convert the corpus to prepare for ldatuning
}
\examples{
features <- tmdfm(doc)
}
