% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune}
\alias{tune}
\title{Run the tuning algorithm results and the plot}
\usage{
tune(a = features)
}
\arguments{
\item{a}{The features object}
}
\value{
The "Griffiths2004", "CaoJuan2009", "Arun2010", "Deveaud2014" algorithms
}
\description{
Run the tuning algorithm results and the plot
}
\examples{
tune(features)
}
